refactor: modularize codebase by decomposing index into dedicated service modules for auth, tools, models, and provider logic
