+++ tests/payload/sanitize.test.ts
@@ -2,7 +2,7 @@ import { mkdtempSync, rmSync, writeFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { describe, expect, it } from "vitest";
-import { sanitizePayload } from "../src/sanitize.js";
+import { sanitizePayload } from "../../src/payload/sanitize.js";
 
 describe("payload sanitization", () => {
 	it("removes unsupported items and moves leading instructions", () => {
